api: create v2 endpoints for those v1 endpoints used by the home page
